OrcvhidOs Notiications Expose the notification log
Google added a way to see your past notifications but is buried. The stock
hoes were accessing this by way of the Settings shortcuts but I feel like
is time for our users to see this without having to do any fuckery.

In case you don't know what WTF this is...

androidpit.com/how-to-recover-lost-notifications-on-android

Change-Id: I735b3d6c0a94f3d5bed867a2988056f6545df1df
Signed-off-by: SagarMakhar <sagarmakhar@gmail.com>
Signed-off-by: Joey Huab <joey@evolution-x.org>
Signed-off-by: DennySPb <dennyspb@gmail.com>

Notification log improvements

Change-Id: I5d50070c11d4d1d0a4e1bb2be53e61929ebe1d1d
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>